refactor(zendesk): isolate mock infrastructure as test-only files
Renamed mock_*.go to mock_*_test.go so Go excludes them from production
builds. Also trimmed mock_scenarios to the single scenario (AuthFailure)
that tests actually use.
